deploy/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quadruped-locomotion-deploy"
version = "0.1.0"
description = "Lightweight deployment for quadruped locomotion RL policies"
requires-python = ">=3.10"
dependencies = [
    "torch",
    "numpy",
    "pyyaml",
]

[project.optional-dependencies]
robot = ["unitree_sdk2py"]
mujoco = ["mujoco"]

[tool.setuptools.packages.find]
include = ["core*", "runners*", "utils*"]

[tool.setuptools.package-data]
"*" = ["*.yaml"]
